Java-1tdwsp

+++ src/Exercicio.java
@@ -18,19 +18,19 @@ public class Exercicio {
 
         //Declarar 3 variaveis(cp1, cp2,cp3)
 
-        double cp1, cp2, cp3;
+        Float cp1, cp2, cp3;
         double media;
 
         // Ver as notas cp
 
         System.out.println("Coloque sua nota do cp1");
-        cp1 = Double.parseDouble (new Scanner(System.in).nextLine());
+        cp1 = Float.parseFloat(new Scanner(System.in).nextLine());
 
         System.out.println("Coloque sua nota do cp2");
-        cp2 = Double.parseDouble(new Scanner (System.in).nextLine());
+        cp2 = Float.parseFloat(new Scanner (System.in).nextLine());
 
         System.out.println("Coloque sua nota do cp3");
-        cp3 = Double.parseDouble(new Scanner (System.in).nextLine());
+        cp3 = Float.parseFloat(new Scanner (System.in).nextLine());
 
 
         media = (cp1 + cp2 +cp3) / 3;
